feat: :sparkles: Set up back end routes for user Events
GetAllEvents now gets user events by mode as well and returns an pbject of all events and the object Ids of the saved user events. Those id's are then checked on the front end against the rest of the events. The handle user events handles the toggling of the save button.